moves header and footer layouts to layouts folder
