--- src/components/Footer.css
@@ -1,7 +1,0 @@
-.main-footer {
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    background-color: lightgray;
-}
