+++ src/index.js
@@ -2,11 +2,9 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Layout from "./components/Layout";
+import Layout from "./layouts/Layout";
 import NoPage from "./components/NoPage";
 import Game from "./components/Game";
-
-
 
 
 function App() {
@@ -20,7 +18,6 @@ function App() {
       </Routes>
     </BrowserRouter>
   )
-
 }
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
