+++ src/layouts/Footer.css
@@ -1,0 +1,7 @@
+.main-footer {
+    background-color: lightgray;
+}
+
+.main-footer a{
+    color: black;
+}
