+++ src/layouts/Header.js
@@ -1,8 +1,6 @@
-import { Link } from "react-router-dom"
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
 
 const Header = () => {
   return (
@@ -13,7 +11,7 @@ const Header = () => {
         <Navbar.Collapse id="basic-navbar-nav">
           <Nav className="me-auto">
             <Nav.Link href="/">Home</Nav.Link>
-            <Nav.Link href="/NoPage">placeholder</Nav.Link>
+            {/* <Nav.Link href="/NoPage">placeholder</Nav.Link> */}
           </Nav>
         </Navbar.Collapse>
       </Container>
